clean up

+++ app/alt_pll_reconfig_interface.h
@@ -53,16 +53,15 @@ class alt_pll_reconfig_interface
  private:
  protected:
   unsigned int m_speed_grade;
-  unsigned int m_fout_min, m_fout_max, m_fout;
-  unsigned int m_fref_min, m_fref_max, m_fref;
-  unsigned int m_fpfd_min, m_fpfd_max, m_fpfd;
-  unsigned int m_fvco_max, m_fvco_min, m_fvco;
-  
-  unsigned int m_c_min, m_c_max, m_c_count, m_c_high, m_c_low;
+  unsigned long long m_fout_min, m_fout_max, m_fout;
+  unsigned long long m_fref_min, m_fref_max, m_fref;
+  unsigned long long m_fpfd_min, m_fpfd_max, m_fpfd;
+  unsigned long long m_fvco_max, m_fvco_min, m_fvco;
+  unsigned long long m_c_min, m_c_max, m_c_count, m_c_high, m_c_low;
   bool m_c_bypass, m_c_odd;
-  unsigned int m_n_min, m_n_max, m_n_count, m_n_high, m_n_low;
+  unsigned long long m_n_min, m_n_max, m_n_count, m_n_high, m_n_low;
   bool m_n_bypass, m_n_odd;
-  unsigned int m_m_min, m_m_max, m_m_count, m_m_high, m_m_low;
+  unsigned long long m_m_min, m_m_max, m_m_count, m_m_high, m_m_low;
   bool m_m_bypass, m_m_odd;
   bool sys_file_exist(const std::string& file);
   std::string m_family;
